+++ NEWS.md
@@ -1,7 +1,7 @@
 # teal.modules.general 0.2.8.9000
+* Added persistence for `lengthMenu` selected value from `datatable` in `tm_data_table`. 
 * Added download, enlarge and resize graph options to `tm_variable_browser` module.
 * Added download and expand options to `tm_t_crosstable`.
-
 Issues with no news:
 
 # teal.modules.general 0.2.8