+++ NEWS.md
@@ -1,3 +1,5 @@
+# teal.modules.general 0.2.9.9000
+
 # teal.modules.general 0.2.9
 ### Enhancements
 * Added download, enlarge and resize graph options to `tm_variable_browser` module.